Set the zeros function's dtype
Former-commit-id: fe38aed4420a3c34729fffab3c791301d06bc9ae

+++ scripts/sct_image.py
@@ -278,7 +278,7 @@ def pad_image(im, pad_x_i=0, pad_x_f=0, pad_y_i=0, pad_y_f=0, pad_z_i=0, pad_z_f
         im.data = im.data.reshape(new_shape)
 
     # initialize padded_data, with same type as im.data
-    padded_data = zeros((nx + pad_x_i + pad_x_f, ny + pad_y_i + pad_y_f, nz + pad_z_i + pad_z_f)).astype(im.data.dtype)
+    padded_data = zeros((nx + pad_x_i + pad_x_f, ny + pad_y_i + pad_y_f, nz + pad_z_i + pad_z_f), dtype=im.data.dtype)
 
     if pad_x_f == 0:
         pad_x_f = None
